extracts update-model functionality to separate file
